Allow compiling and running fuzzing tests as regular Nim programs
When you compile a fuzzing test with `nim c -r test.nim`, it will
now compile to a normal program accepting a fuzzing input file as
a command-line argument. This is an useful way to debug crashes
discovered by libFuzzer or AFL.

The code was also refactored to reduce the number of define variables
affecting the build type (there is no longer a separate -d:afl flag).