Add flag to `generateReport` to allow disabling total in Markdown report
The lines that indicate total number of tests in Markdown reports can
lead to merge conflicts if the report is committed into the repo as is
done in `status-im/nimbus-eth2`. Fixing the merge conflicts can be time
intensive as tests may have to be re-run to obtain proper counts. Adding
an option to disable emission of totals ensures that this hurdle for
extending the test suite is gone.